+++ Changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - bug/370-anthropic-streaming (2025-09-16)
 
 ### Changed
+  - feat/382-move-threadid-gen-to-backend (2025-10-01)
   - feat/380-subagent-configure (2025-09-29)
   - feat/387-ip-rate-limit (2025-09-28)
   - feat/381-delete-agent (2025-09-28)